Reading eventlog

read_eventlog reads eventlog stored in a file or in a database.

read_eventlog(file = NULL, db = NULL, last_n = Inf, verbose = TRUE)

Arguments

file
A character string. Path to a file log.

db
A character string. Connection string to a mongo database.

last_n
An integer. How many last event records should be return? Default is Inf which returns the whole eventlog.

verbose
A logical value. Should the function print addition messages? Default is TRUE.

Value

An object of class eventlog which is a data frame with appropriate case, activity and timestamp classifiers specified. The eventlog object is a result of bupaR:eventlog function from bupaR package and it is suitable for further process-mining analysis.

See also

Examples

read_eventlog(
 last_n = 25,
 file = system.file("shiny", "demoapp/events.log",
                     package = "shinyEventLogger"))

#> Reading log file: D:/_users/_docs/# Rscripts/shinyEventLogger/inst/shiny/demoapp/events.log

#> The last event is at the top.

#> Event log consisting of:
#> 25 events
#> 2 traces
#> 2 cases
#> 14 activities
#> 23 activity instances
#> 
#> # A tibble: 25 x 17
#>    event_counter event_type event_name event_status session_id
#>    <chr>         <chr>      <chr>      <fct>        <chr>     
#>  1 #12           TEST       testthat:~ SUCCESS      55ae1cf97~
#>  2 #11           EVENT      Number of~ FIRED        55ae1cf97~
#>  3 #10           VALUE      input$bins FIRED        55ae1cf97~
#>  4 #9            VALUE      input$var~ FIRED        55ae1cf97~
#>  5 #8            EVENT      Variable ~ FIRED        55ae1cf97~
#>  6 #7            OUTPUT     head(data~ FIRED        55ae1cf97~
#>  7 #6            OUTPUT     str(datas~ FIRED        55ae1cf97~
#>  8 #5            VALUE      NROW(data~ FIRED        55ae1cf97~
#>  9 #4            EVENT      Loading d~ DONE         55ae1cf97~
#> 10 #4            EVENT      Loading d~ STARTED      55ae1cf97~
#> # ... with 15 more rows, and 12 more variables: event_timestamp <dttm>,
#> #   event_body <chr>, event_id <chr>, dataset <chr>, fun <chr>, resource <fct>,
#> #   build <int>, secs <dbl>, n_rows <int>, bins <int>, variable <chr>,
#> #   .order <int>