Reading eventlog

read_eventlog reads eventlog stored in a file or in a database.

read_eventlog(file = NULL, db = NULL, last_n = Inf, verbose = TRUE)

Arguments

file
A character string. Path to a file log.

db
A character string. Connection string to a mongo database.

last_n
An integer. How many last event records should be return? Default is Inf which returns the whole eventlog.

verbose
A logical value. Should the function print addition messages? Default is TRUE.

Value

An object of class eventlog which is a data frame with appropriate case, activity and timestamp classifiers specified. The eventlog object is a result of bupaR:eventlog function from bupaR package and it is suitable for further process-mining analysis.

See also

Examples

# NOT RUN {
read_eventlog(
 last_n = 25,
 file = system.file("shiny", "demoapp/events.log",
                     package = "shinyEventLogger"))
# }